+++ CHANGELOG.md
@@ -9,3 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - plugin init based on fake data
+
+### Changed
+- plugin config dumping